requirements.txt
python-docx
langchain_openai
aiogram==2.11